+++ .gitignore
@@ -5,5 +5,6 @@ riderModule.iml
 /_ReSharper.Caches/
 .idea/
 /.vs
+/Traveller.Tools/Traveller.Tools.sln.DotSettings.user
 /Traveler.Api/Properties/launchSettings.json
 /Traveller.Api/Properties/launchSettings.json
