+++ Traveller.Tools/Traveller.Tools.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Traveller.Tools", "Traveller.Tools.csproj", "{A7D15633-2BE0-478E-AF77-25072CCF5778}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A7D15633-2BE0-478E-AF77-25072CCF5778}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A7D15633-2BE0-478E-AF77-25072CCF5778}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A7D15633-2BE0-478E-AF77-25072CCF5778}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A7D15633-2BE0-478E-AF77-25072CCF5778}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
